refactor : interceptor를 filter로 변경
security를 적용하고 인증 필터가 동작하기 전에 작성했던 interceptor를 동작시키기 위해
interceptor를 filter로 변경, 테스트 확인!

issue key : SB2-11
