+++ simpleBoard02/server-b-user/src/main/resources/application.yml
